pyproject.toml
[tool.poetry]
name = "cardanoapi"
version = "1.0.1"
description = "API to interact with cardanopythonlib"
authors = ["Moxie <luis.restrepo@ayllu.io>"]
readme = "README.md"
license = "Apache-2.0"
repository = "https://github.com/larestrepo/CardanoPythonLib"
keywords = ["cardano", "blockchain", "python", "backend"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.19.0"
fastapi = "^0.85.1"
psycopg2 = "^2.9.4"
cardanopythonlib = "^1.0.5"
python-multipart = "^0.0.5"
celery = "^5.2.7"
redis = "^4.3.4"
flower = "^1.2.0"
sqlalchemy = "^1.4.42"
alembic = "^1.8.1"
sqlalchemy-utils = "^0.38.3"
gunicorn = "^20.1.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
